LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/private-network-access/secure-context.https.window-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'treat-as-public-address'.


PASS Local secure page fetches local page.
PASS Public secure page fetches local page.
